Add test helpers instead of importing from test file directly
This avoids running tests twice.
